Move and Rename SessionSocketController
